Added new checks to see if user is in role.

+++ StudyPlatform/Controllers/CourseController.cs
@@ -7,6 +7,9 @@ using StudyPlatform.Services.Users.Interfaces;
 using StudyPlatform.Infrastructure;
 using static StudyPlatform.Common.GeneralConstants;
 using StudyPlatform.Infrastructure.Infrastructure;
+using StudyPlatform.Data.Models;
+using Microsoft.AspNetCore.Identity;
+using static StudyPlatform.Common.GeneralConstants;
 
 namespace StudyPlatform.Controllers
 {
@@ -17,15 +20,18 @@ namespace StudyPlatform.Controllers
         private readonly ICourseViewService _courseViewService;
         private readonly ICategoryViewService _categoryViewService;
         private readonly ICourseViewFormService _courseViewFormService;
+        private readonly UserManager<ApplicationUser> _userManager;
 
         public CourseController(
             ICourseViewService courseViewService,
             ICategoryViewService categoryViewService,
-            ICourseViewFormService courseViewFormService)
+            ICourseViewFormService courseViewFormService,
+            UserManager<ApplicationUser> userManager)
         {
             this._courseViewService = courseViewService;
             this._categoryViewService = categoryViewService;
             this._courseViewFormService = courseViewFormService;
+            this._userManager = userManager;
         }
 
         /// <summary>
@@ -57,10 +63,17 @@ namespace StudyPlatform.Controllers
         /// <param name="courseId"></param>
         /// <returns></returns>
         [AutoValidateAntiforgeryToken]
-        [Authorize(Roles = $"{TeacherRoleName},{AdministratorRoleName}")]
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> Edit(int courseId)
         {
+            ApplicationUser appUser = await this._userManager.GetUserAsync(User);
+            bool isTeacher = await this._userManager.IsInRoleAsync(appUser, TeacherRoleName);
+            if (!isTeacher)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!await this._courseViewService.AnyByIdAsync(courseId))
             {
                 return BadRequest();
@@ -77,10 +90,17 @@ namespace StudyPlatform.Controllers
         /// <param name="model"></param>
         /// <returns></returns>
         [AutoValidateAntiforgeryToken]
-        [Authorize(Roles = $"{TeacherRoleName},{AdministratorRoleName}")]
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Edit(CourseViewFormModel model)
         {
+            ApplicationUser appUser = await this._userManager.GetUserAsync(User);
+            bool isTeacher = await this._userManager.IsInRoleAsync(appUser, TeacherRoleName);
+            if (!isTeacher)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!ModelState.IsValid){
                 return View();
             }
@@ -96,10 +116,17 @@ namespace StudyPlatform.Controllers
         /// <param name="courseId"></param>
         /// <returns></returns>
         [AutoValidateAntiforgeryToken]
-        [Authorize(Roles = $"{TeacherRoleName},{AdministratorRoleName}")]
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> Remove(int courseId)
         {
+            ApplicationUser appUser = await this._userManager.GetUserAsync(User);
+            bool isTeacher = await this._userManager.IsInRoleAsync(appUser, TeacherRoleName);
+            if (!isTeacher)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             int categoryId = await this._categoryViewService.GetCategoryIdByCourseIdAsync(courseId);
             string categoryName = await this._categoryViewService.GetNameUrlByIdAsync(categoryId);
             await this._courseViewFormService.RemoveAsync(courseId);
@@ -113,10 +140,17 @@ namespace StudyPlatform.Controllers
         /// <param name="categoryId"></param>
         /// <returns></returns>
         [AutoValidateAntiforgeryToken]
-        [Authorize(Roles = $"{TeacherRoleName},{AdministratorRoleName}")]
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> Add(int categoryId)
         {
+            ApplicationUser appUser = await this._userManager.GetUserAsync(User);
+            bool isTeacher = await this._userManager.IsInRoleAsync(appUser, TeacherRoleName);
+            if (!isTeacher)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!await this._categoryViewService.AnyByIdAsync(categoryId))
             {
                 return BadRequest();
@@ -135,11 +169,18 @@ namespace StudyPlatform.Controllers
         /// </summary>
         /// <param name="model"></param>
         /// <returns></returns>
-        [Authorize(Roles = $"{TeacherRoleName},{AdministratorRoleName}")]
+        [Authorize]
         [AutoValidateAntiforgeryToken]
         [HttpPost]
         public async Task<IActionResult> Add(CourseViewFormModel model)
         {
+            ApplicationUser appUser = await this._userManager.GetUserAsync(User);
+            bool isTeacher = await this._userManager.IsInRoleAsync(appUser, TeacherRoleName);
+            if (!isTeacher)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!ModelState.IsValid)
             {
                 return View();
